# contracts-gen/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangeStatusRequest": {
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
      "CreateFlagRequest": {
        "properties": {
          "defaultOutcome": {
            "type": "boolean"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "rollout": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/RolloutDto"
              }
            ],
            "type": "object"
          },
          "targetingRules": {
            "items": {
              "$ref": "#/components/schemas/TargetingRuleDto"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "required": [
          "key",
          "name",
          "defaultOutcome",
          "targetingRules",
          "rollout"
        ],
        "type": "object"
      },
...
      "FlagResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "defaultOutcome": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "killSwitchEngaged": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "rollout": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/RolloutDto"
              }
            ],
            "type": "object"
          },
          "status": {
            "type": "string"
          },
          "targetingRules": {
            "items": {
              "$ref": "#/components/schemas/TargetingRuleDto"
            },
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "key",
          "name",
          "defaultOutcome",
          "status",
          "killSwitchEngaged",
          "version",
          "targetingRules",
          "rollout",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "KillSwitchRequest": {
        "properties": {
          "engaged": {
            "type": "boolean"
          }
        },
        "required": [
          "engaged"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "RolloutDto": {
        "properties": {
          "percentage": {
            "format": "int32",
            "type": "integer"
          },
          "salt": {
            "type": "string"
          }
        },
        "required": [
          "percentage",
          "salt"
        ],
        "type": "object"
      },
...
      "TargetingRuleDto": {
        "properties": {
          "attribute": {
            "type": "string"
          },
          "operator": {
            "type": "string"
          },
          "outcome": {
            "type": "boolean"
          },
          "values": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "attribute",
          "operator",
          "values",
          "outcome"
        ],
        "type": "object"
      },
...
      "UpdateFlagRequest": {
        "properties": {
          "defaultOutcome": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "rollout": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/RolloutDto"
              }
            ],
            "type": "object"
          },
          "targetingRules": {
            "items": {
              "$ref": "#/components/schemas/TargetingRuleDto"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "required": [
          "name",
          "defaultOutcome",
          "targetingRules",
          "rollout"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OptiPulse.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.0.4",
  "paths": {
...
    "/api/v1/flags": {
      "get": {
        "operationId": "ListFlags",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FlagResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Flags"
        ]
      },
      "post": {
        "operationId": "CreateFlag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFlagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagResponse"
                }
              }
            },
            "description": "Created"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "tags": [
          "Flags"
        ]
      }
    },
    "/api/v1/flags/{key}": {
      "get": {
        "operationId": "GetFlag",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Flags"
        ]
      },
      "put": {
        "operationId": "UpdateFlag",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFlagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Precondition Required"
          }
        },
        "tags": [
          "Flags"
        ]
      }
    },
    "/api/v1/flags/{key}/kill-switch": {
      "post": {
        "operationId": "SetKillSwitch",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Flags"
        ]
      }
    },
    "/api/v1/flags/{key}/status": {
      "post": {
        "operationId": "ChangeFlagStatus",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "tags": [
          "Flags"
        ]
      }
    },
...
  },
...
}